refactor: 화면 컴포넌트 디렉토리명 변경
