+++ src/features/misc/index.ts
@@ -1,1 +1,1 @@
-export * from './routes';
+export * from './screens';
