+++ src/features/misc/screens/SplashScreen/index.ts
