+++ src/features/quests/index.ts
@@ -1,3 +1,3 @@
 export * from './types';
 
-export * from './routes';
+export * from './screens';
