+++ src/features/quests/screens/QuestDetailScreen/QuestDetailScreen.tsx
